+++ src/blosc2/__init__.py
@@ -567,6 +567,7 @@ from .dict_store import DictStore
 from .tree_store import TreeStore
 from .batch_array import Batch, BatchArray
 from .list_array import ListArray
+from .utf8_array import Utf8Array, utf8_array
 from .objectarray import ObjectArray, objectarray_from_cframe
 from .ref import Ref
 from .b2objects import open_b2object
@@ -832,6 +833,7 @@ __all__ = [  # noqa : RUF022
     "uint32",
     "uint64",
     "utf8",
+    "utf8_array",
     "vlbytes",
     "vlstring",
     # Grouped reductions
@@ -879,6 +881,7 @@ __all__ = [  # noqa : RUF022
     "Tuner",
     "URLPath",
     "ObjectArray",
+    "Utf8Array",
     # Version
     "__version__",
     # Utils